added ability to read from page without going through entire read process
simplefs kernel page read does the kernel read, but it doesn't jump through all of the
steps that the normal kernel read does, it sets changes the fs register and 
then calls copy_page_to_iter on the passed in page, and writes to the buffer that
was passed in. This also takes a count, and an offset thing. If we are just
reading a page, then we should probably just have the count be the page size
and the offset be zero.
 
There is a function "invalidatepage" which takes an inode number, a page number
and a buffer. The purpose of this function is that if the page exists in
the page cache, it will write the data from the page into the passed in buffer.